add img_file check

+++ db/leancloud_upload.py
@@ -112,11 +112,17 @@ def exist_file(filename):
         return False
 
 
+def is_img_file(suffix):
+    img_types = set(['jpg', 'png', 'gif', 'jpeg'])
+    return suffix in img_types
+
+
 def upload_all_file(file_dir):
     file_list = get_file_list(file_dir)
     for each_file in file_list:
         filename = os.path.basename(each_file)    # have suffix
-        if not exist_file(filename):
+        suffix = each_file.split('.')[-1]
+        if not exist_file(filename) and is_img_file(suffix):
             upload_file(each_file)
             time.sleep(3)
 
@@ -133,4 +139,4 @@ def main():
 
 if __name__ == '__main__':
     main()
-    #test()
+    # test()
